[hotfix] Factor ExecutionGraph creation out into ExecutionGraphFactory
Using the ExecutionGraphFactory for creating and restoring an ExecutionGraph allows
to share the functionality between the DefaultScheduler and the AdaptiveScheduler.
